requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
httpx==0.28.0

google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0

pydantic-settings==2.10.1
loguru==0.7.3
tenacity==9.1.2

# testing & tooling
pytest==8.3.2
ruff==0.12.8

pdfminer.six==20231228